Moved pubsub message processing into rows into a different class More cleanup and refactoring
