+++ clients/venice-push-job/src/main/java/com/linkedin/venice/spark/SparkConstants.java
@@ -52,15 +52,14 @@ public class SparkConstants {
 
   public static final StructType BASIC_RAW_PUBSUB_INPUT_TABLE_SCHEMA = new StructType(
       new StructField[] { new StructField(KEY_COLUMN_NAME, BinaryType, false, Metadata.empty()), // serialized key
-          new StructField(VALUE_COLUMN_NAME, BinaryType, true, Metadata.empty()), // serialized value
-          new StructField("__partition__", IntegerType, false, Metadata.empty()),
-          new StructField("__message_type__", IntegerType, false, Metadata.empty()), // enum of put/delete/update
-          new StructField("__offset__", LongType, true, Metadata.empty()), // offset in the topic
           new StructField("__region__", StringType, false, Metadata.empty()),
+          new StructField("__partition__", IntegerType, false, Metadata.empty()),
+          new StructField("__offset__", LongType, true, Metadata.empty()), // offset in the topic
+          new StructField("__message_type__", IntegerType, false, Metadata.empty()), // enum of put/delete/update
+          new StructField("__schema_id__", IntegerType, false, Metadata.empty()),
           new StructField(KEY_COLUMN_NAME, BinaryType, false, Metadata.empty()),
           new StructField(VALUE_COLUMN_NAME, BinaryType, true, Metadata.empty()),
 
-      // new StructField("__schema_id__", IntegerType, false, Metadata.empty()),
       // new StructField("__replication_metadata_version_id__", IntegerType, false, Metadata.empty()),
       // new StructField("__replication_metadata_payload__", BinaryType, false, Metadata.empty()) });
       });
