+++ clients/venice-push-job/src/main/java/com/linkedin/venice/spark/input/pubsub/raw/VeniceRawPubsubInputScan.java
@@ -1,4 +1,4 @@
-package com.linkedin.venice.spark.input.table.rawPubSub;
+package com.linkedin.venice.spark.input.pubsub.raw;
 
 import static com.linkedin.venice.vpj.VenicePushJobConstants.*;
